move logic to search module
