+++ app-server/src/routes/mod.rs
@@ -3,7 +3,6 @@ pub mod probes;
 pub mod realtime;
 pub mod rollouts;
 pub mod signals;
-pub mod snippets;
 pub mod spans;
 pub mod sql;
 pub mod types;
